Moved IMGUI code to NetworkManagerGUI
