+++ Assets/Mirage/Runtime/NetworkManagerGUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 81da5905cb362874ba7cbe36247debba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 7453abfe9e8b2c04a8a47eb536fe21eb, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
